+++ src/tests/test2/vk/MultithreadedTerrainSceneTest.cpp
@@ -393,7 +393,7 @@ void MultithreadedTerrainSceneTest::prepareSecondaryCommandBuffer(std::size_t th
         {
             auto indexSize = sizeof(terrain().indices().front());
             auto renderChunk = [&cmdBuffer, indexSize](std::size_t count, std::ptrdiff_t offset) {
-                cmdBuffer.drawIndexed(count, 1, offset / indexSize, 0, 0);
+                cmdBuffer.drawIndexed(static_cast<uint32_t>(count), 1, static_cast<uint32_t>(offset / indexSize), 0, 0);
             };
             // TODO: think on how to extend this beyond 4-threads
             terrain().executeLoD(currentPosition(), renderChunk, threadIndex);
