+++ src/tests/test3/vk/MultithreadedShadowMappingSceneTest.cpp
@@ -737,7 +737,7 @@ void MultithreadedShadowMappingSceneTest::prepareSecondaryCommandBuffer(std::siz
             cmdBuffer.pushConstants(pass.pipelineLayout, vk::ShaderStageFlagBits::eVertex, 0, sizeof(MVP), &MVP);
 
             cmdBuffer.bindVertexBuffers(0, {{renderObject.vbo.buffer}}, {{0}});
-            cmdBuffer.draw(renderObject.drawCount, 1, 0, 0);
+            cmdBuffer.draw(static_cast<uint32_t>(renderObject.drawCount), 1, 0, 0);
         }
 
         cmdBuffer.end();
@@ -770,7 +770,7 @@ void MultithreadedShadowMappingSceneTest::prepareSecondaryCommandBuffer(std::siz
                                     &matrices);
 
             cmdBuffer.bindVertexBuffers(0, {{renderObject.vbo.buffer}}, {{0}});
-            cmdBuffer.draw(renderObject.drawCount, 1, 0, 0);
+            cmdBuffer.draw(static_cast<uint32_t>(renderObject.drawCount), 1, 0, 0);
         }
 
         cmdBuffer.end();
